Made Makefile considerably smarter.
Looks for any folders in "libraries" without children.
Each folder is compiled into a separate .so
If a LinkDef.h is present, a Dictionary will be generated and added to the .so.
Each src file in the directories "utils" and "sandbox" will be made into an executable.

+++ include/Unpacker.hh
@@ -26,10 +26,9 @@ private:
 
   bool uses_fragment_header;
   std::ifstream infile;
-  long bytes_read;
+  size_t bytes_read;
   size_t total_size;
-  long items_unpacked;
-
+  size_t items_unpacked;
 };
 
 #endif
